move printing service to the preview chunk
