+++ src/app/testing/app-testing.module.ts
@@ -60,7 +60,7 @@ import { NodeActionsService } from '../services/node-actions.service';
 import { NodePermissionService } from '../services/node-permission.service';
 import { ContentApiService } from '../services/content-api.service';
 import { ExtensionService } from '../extensions/extension.service';
-import {ViewUtilService} from '../services/view-util.service';
+import { ViewUtilService } from '../components/preview/view-util.service';
 
 @NgModule({
     imports: [
